Fixes for Cabal >= 3.0.
Standard package management repositories for different operating systems
do not currently contain older versions of cabal-install.

Newer versions of cabal make it more convenient to install a package
if there is a cabal.project file declaring all the cabal files included
in all libraries.

However, newer versions of cabal do not install BNFC as an executable in
the PATH or in the $PWD/.cabal-sandbox/bin directory, instead picking
BNFC as a library if declared in the build-depends of the custom-setup
section in the cabal file (as it is now).

This commit introduces several changes:
- Add a cabal.project file that declares all cabal files in the repo.
- Make cabal 3.0 the version of all cabal files.
- Move all cabal-version: declarations to the top of the file
as required by the new format.
- Introduce a new section build-tool-depends to all sections that depend
on BNFC as a build tool, using BNFC:bnfc to depend specifically on the
executable (removed from the custom-setup section).
- Change the license to NASA-1.3 as required by the cabal 3.0 format.