ogma-core: Use template expansion to generate F' monitoring component. Refs #185.
The F' backend uses a fixed template to generate the F' component. That
template does not fit all use cases, so we are finding users heavily
modifying the output (which is hard to keep up with when there are
changes), and or not using ogma altogether for that reason.

This commit modifies the ogma-core fprime command to use mustache to
generate the F' monitoring application via a template and variable
expansion. We adapt the template to also use those variables, and we
modify the cabal file to include the new files as data files that are
copied over during installation.