ogma-cli: Add auxiliary files used by tests to Cabal package. Refs #216.
The cabal file for ogma-cli is not including all the files needed for
the tests, causing the tests to fail when the packages are grabbed from
hackage rather than the git repo.

A prior commit has modified the tests so that the auxiliary files needed
by the tests are packaged with ogma-cli.

This commit adds all examples mentioned by the tests to the Cabal
package's data-files field.