ogma-cli: Modify tests to file files from package installation data files. Refs #216.
The cabal file for ogma-cli is not including all the files needed for
the tests, causing the tests to fail when the packages are grabbed from
hackage rather than the git repo.

This commit uses the functions generated by Cabal to locate those files
from the installation directory during testing, so that they can be
packaged with ogma-cli. The new implementation needs access to functions
from the `filepath` package, so it is added to the corresponding section
in the Cabal package as a dependency of the tests.

The Cabal file is modified to include the file autogenerated by Cabal
used to help locate data files in the test's other-modules field.