ogma-language-c: Add auxiliary files used by tests to Cabal package.
The cabal file for ogma-language-c is not including all the files needed
for the tests, causing the tests to fail when the packages are grabbed
from hackage rather than the git repo.

A prior commit has modified the tests so that the auxiliary files needed
by the tests are packaged with ogma-language-c.

This commit adds all examples mentioned by the tests to the Cabal
package's data-files field.