ogma-core: Introduce search command. Refs #464.
As Ogma becomes used with larger projects, it is useful to add a search
command shows a list of all requirements and models that match a given
query.

This commit adds a new Search command to `ogma-core` that find a given
query in the input files and produces a list with the requirements or
diagrams that match the search query.