Add sample image of runnning generated ROS 2 application.
To facilitate adoption of Ogma, we want to have a collection of examples
that demonstrate the most common use cases. A demonstration of how to
use the ROS 2 backend to generate a minimal application would be a good
starting point for users that want to generate monitors for ROS 2.

This commit adds an image that illustrates how an Ogma-generated ROS 2
application can be executed and tested.