ogma-cli: Add CI job to test diagram backend. Refs #332.
To keep Ogma working in future versions and ensure that the changes
introduced do not affect the code generated from diagrams in a negative
way, we want to have a test that checks that the code generated at least
compiles.

This commit add a new CI action file that generates a Copilot
specification from a state machine diagram, compiles it to C, and links
it with a main driver that executes a run with predictable transitions.
The result is compared with an expectation. If the result and the
expectation do not match, the CI job exits with a non-zero exit code.

This action is based on the first diagram example under
ogma-cli/examples.